[FIX] Make ExpireReservation job consider reservation's status
* Purpose
** What is the objective?
To make sure that the ExpireReservation job only deletes on-hold
reservations.

** Are we solving a problem or making necessary changes to support an
Solving a problem. Previously, even a found reservation that's confirmed
will be deleted.

** Please provide any additional context that may
    help us understand the changes.
None.

* Summary
** Are there files we can ignore?
None.

** Please provide a high-level summary of the changes.
- Adds check on reservation status in ExpireReservation.php
- Adds test in ExpireReservationTest.php
  - Validates confirmed reservation isn't deleted

+++ app/Jobs/ExpireReservation.php
@@ -2,6 +2,7 @@
 
 namespace App\Jobs;
 
+use App\Enums\ReservationStatus;
 use App\Models\Reservation;
 use DB;
 use Illuminate\Contracts\Queue\ShouldQueue;
@@ -20,7 +21,7 @@ class ExpireReservation implements ShouldQueue
     {
         DB::transaction(function () {
             $reservation = Reservation::lockForUpdate()->find($this->reservation_id);
-            if ($reservation) {
+            if ($reservation && $reservation->status === ReservationStatus::OnHold) {
                 $reservation->delete();
             }
         });
